1) Add print for StatusCode 2) Fix some pedantic compilator warnings

+++ source/src/format.cpp
@@ -22,7 +22,7 @@ std::string Ostap::format ( const std::string& fmt    ,
   static char s_buffer[ s_LEN ] ;
   const int result = snprintf ( s_buffer , s_LEN , fmt.c_str() , value1 ) ;
   return 
-    0 <= result && result < s_LEN ?  
+    0 <= result && (unsigned int) result < s_LEN ?  
     std::string ( s_buffer , s_buffer + result ) : 
     fmt + std::to_string ( value1 ) ;
 }
@@ -38,7 +38,7 @@ std::string Ostap::format ( const std::string& fmt    ,
   const int result = snprintf ( s_buffer , s_LEN , fmt.c_str() , 
                                 value1 , value2 ) ;
   return 
-    0 < result && result < s_LEN ?  
+    0 < result && (unsigned int) result < s_LEN ?  
     std::string ( s_buffer , s_buffer + result ) : 
     fmt + std::to_string ( value1 ) + 
     " " + std::to_string ( value2 ) ;
@@ -56,7 +56,7 @@ std::string Ostap::format ( const std::string& fmt    ,
   const int result = snprintf ( s_buffer , s_LEN , fmt.c_str() , 
                                 value1 , value2 , value3 ) ;
   return 
-    0 < result && result < s_LEN ? std::string ( s_buffer , s_buffer + result ) : 
+    0 < result && (unsigned int) result < s_LEN ? std::string ( s_buffer , s_buffer + result ) : 
     ( fmt + std::to_string ( value1 ) + 
       " " + std::to_string ( value2 ) ,
       " " + std::to_string ( value3 ) ) ;
@@ -76,7 +76,7 @@ std::string Ostap::format ( const std::string& fmt    ,
                                 value1 , value2 , value3 , value4 ) ;
   //
   return 
-    0 < result && result < s_LEN ? std::string ( s_buffer , s_buffer + result ) : 
+    0 < result && (unsigned int) result < s_LEN ? std::string ( s_buffer , s_buffer + result ) : 
     ( fmt + std::to_string ( value1 ) + 
       " " + std::to_string ( value2 ) ,
       " " + std::to_string ( value3 ) , 
@@ -98,7 +98,7 @@ std::string Ostap::format ( const std::string& fmt    ,
                                 value1   , value2 , value3 , value4 , value5 ) ;
   //
   return 
-    0 < result && result < s_LEN ? std::string ( s_buffer , s_buffer + result ) :
+    0 < result && ( unsigned int) result < s_LEN ? std::string ( s_buffer , s_buffer + result ) :
     ( fmt + std::to_string ( value1 ) + 
       " " + std::to_string ( value2 ) ,
       " " + std::to_string ( value3 ) ,
